Use pinouts from the datasheet for GPIO ports
There's conflicting information about what the pins are named and where
they live on the various GPIO ports. The Ameba Arduino SDK `PinNames.h`
makes it clear-as-mud that the pin names like PA1 correspond to GPIO
port A pin 1, and that those should match with pins labeled like GA1 on
the RTL-00 module.